refactor(SelectDevice): rework components DeviceItem with Item + wording
