+++ src/components/SelectDevice/index.js
@@ -8,14 +8,15 @@ import { useNavigation, useTheme } from "@react-navigation/native";
 import { discoverDevices } from "@ledgerhq/live-common/lib/hw";
 import type { TransportModule } from "@ledgerhq/live-common/lib/hw";
 import type { Device } from "@ledgerhq/live-common/lib/hw/actions/types";
+import Icon from "react-native-vector-icons/dist/Feather";
 import { ScreenName } from "../../const";
 import { knownDevicesSelector } from "../../reducers/ble";
-import DeviceItem from "../DeviceItem";
+import DeviceItem from "./DeviceItem";
 import BluetoothEmpty from "./BluetoothEmpty";
 import USBEmpty from "./USBEmpty";
 import LText from "../LText";
 import Animation from "../Animation";
-import PairNewDeviceButton from "./PairNewDeviceButton";
+import Item from "./Item";
 
 import lottieUsb from "../../screens/Onboarding/assets/nanoS/plugDevice/data.json";
 import { track } from "../../analytics";
@@ -23,6 +24,7 @@ import { track } from "../../analytics";
 type Props = {
   onBluetoothDeviceAction?: (device: Device) => void,
   onSelect: (device: Device) => void,
+  onWithoutDevice?: () => void,
   withArrows?: boolean,
   usbOnly?: boolean,
   filter?: (transportModule: TransportModule) => boolean,
@@ -34,6 +36,7 @@ export default function SelectDevice({
   withArrows,
   filter = () => true,
   onSelect,
+  onWithoutDevice,
   onBluetoothDeviceAction,
   autoSelectOnAdd,
 }: Props) {
@@ -124,7 +127,13 @@ export default function SelectDevice({
         <View>
           <BluetoothHeader />
           {ble.map(renderItem)}
-          <PairNewDeviceButton onPress={onPairNewDevice} />
+          <Item
+            primary
+            onPress={onPairNewDevice}
+            event="AddDevice"
+            icon={<Icon name="plus" color={colors.live} size={16} />}
+            title={<Trans i18nKey="SelectDevice.deviceNotFoundPairNewDevice" />}
+          />
         </View>
       )}
       {hasUSBSection &&
@@ -139,22 +148,42 @@ export default function SelectDevice({
       ) : (
         other.map(renderItem)
       )}
+      {onWithoutDevice && (
+        <View>
+          <WithoutDeviceHeader />
+          <Item
+            primary
+            withArrow
+            onPress={onWithoutDevice}
+            event="WithoutDevice"
+            title={<Trans i18nKey="SelectDevice.withoutDevice" />}
+          />
+        </View>
+      )}
     </>
   );
 }
 
 const BluetoothHeader = () => (
-  <View style={styles.bluetoothHeader}>
-    <LText semiBold style={styles.section} color="grey">
+  <View style={styles.header}>
+    <LText style={styles.headerText} color="grey">
       <Trans i18nKey="common.bluetooth" />
     </LText>
   </View>
 );
 
 const USBHeader = () => (
-  <LText semiBold style={styles.section} color="grey">
+  <LText style={styles.headerText} color="grey">
     <Trans i18nKey="common.usb" />
   </LText>
+);
+
+const WithoutDeviceHeader = () => (
+  <View style={styles.header}>
+    <LText style={styles.headerText} color="grey">
+      <Trans i18nKey="SelectDevice.withoutDeviceHeader" />
+    </LText>
+  </View>
 );
 
 // Fixme Use the illustration instead of the png
@@ -177,7 +206,12 @@ function getAll({ knownDevices }, { devices }): Device[] {
 }
 
 const styles = StyleSheet.create({
-  section: {
+  header: {
+    flexDirection: "row",
+    justifyContent: "space-between",
+    alignItems: "flex-start",
+  },
+  headerText: {
     fontSize: 14,
     lineHeight: 21,
     marginBottom: 12,
@@ -186,18 +220,6 @@ const styles = StyleSheet.create({
     width: "100%",
     height: 1,
     marginVertical: 24,
-  },
-  addContainer: {
-    flexDirection: "row",
-    alignItems: "center",
-  },
-  bluetoothHeader: {
-    flexDirection: "row",
-    justifyContent: "space-between",
-    alignItems: "flex-start",
-  },
-  or: {
-    marginVertical: 30,
   },
   imageContainer: {
     minHeight: 200,
