+++ src/locales/en/common.json
@@ -2377,10 +2377,12 @@
       "title": "Connect via Bluetooth...",
       "label": "Detect your Nano automatically"
     },
-    "deviceNotFoundPairNewDevice": "Pair via Bluetooth",
+    "deviceNotFoundPairNewDevice": "Add new Ledger Nano X",
     "headerDescription": "Please make sure your {{productName}} is unlocked with Bluetooth enabled",
     "usb": "... or connect USB cable",
     "usbLabel": "Connect USB cable and enter your PIN Code on your device",
+    "withoutDeviceHeader": "I don't have my device with me",
+    "withoutDevice": "Continue without my device",
     "steps": {
       "connecting": {
         "title": "Connecting {{deviceName}}",
